+++ TeleportModule/src/main/java/net/giuse/teleportmodule/database/spawnquery/SaveQuerySpawn.java
@@ -24,6 +24,8 @@ public class SaveQuerySpawn implements Query {
 
     @Override
     public void query() {
+        if(spawnModule.getSpawnBuilderSerializer() == null) return;
+
 
         executeQuery.execute("DROP TABLE Spawn;");
